backend-ml/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dask==2025.2.0
distributed==2025.2.0
duckdb==1.2.1
exceptiongroup==1.2.2
fastapi==0.115.11
fonttools==4.56.0
fsspec==2025.3.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.1.0
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
shap==0.47.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
tblib==3.0.0
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
upstash-redis==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
xgboost==3.0.0
zict==3.0.0
zipp==3.21.0
